+++ cosmos_curate/models/vllm_model_ids.py
@@ -33,7 +33,9 @@ _VLLM_MODELS = {
     "qwen": "Qwen/Qwen2.5-VL-7B-Instruct",
     "nemotron": "nvidia/NVIDIA-Nemotron-Nano-12B-v2-VL-BF16",
     "qwen3_vl_235b": "Qwen/Qwen3-VL-235B-A22B-Instruct",
+    "qwen3_vl_235b_fp8": "Qwen/Qwen3-VL-235B-A22B-Instruct-FP8",
     "qwen3_vl_30b": "Qwen/Qwen3-VL-30B-A3B-Instruct",
+    "qwen3_vl_30b_fp8": "Qwen/Qwen3-VL-30B-A3B-Instruct-FP8",
 }